change to be able to set rateLimit by route directly from fastify
